added validation

+++ Furiax.ExerciseTracker/ExerciseTracker/ExerciseService.cs
@@ -18,31 +18,63 @@ namespace ExerciseTracker
 		internal void DeleteExercise()
 		{
 			var exercises = _exerciseRepository.GetAll().ToList();
-			int id = ExerciseController.GetIdOption(exercises);
-			_exerciseRepository.Delete(id);
+			if(exercises.Count() == 0)
+			{
+				Console.WriteLine("The database is empty");
+				Console.ReadKey();
+			}
+            else
+            {
+				int id = ExerciseController.GetIdOption(exercises);
+				_exerciseRepository.Delete(id);
+			}
 		}
 
 		internal void GetAll()
 		{
 			var exercises = _exerciseRepository.GetAll().ToList();
-			ExerciseController.PrintExercisesTable(exercises);
+			if (exercises.Count == 0)
+			{
+				Console.WriteLine("The database is empty");
+				Console.ReadKey();
+			}
+			else
+			{
+				ExerciseController.PrintExercisesTable(exercises);
+			}
 		}
 
 		internal void GetExerciseById()
 		{
 			var exercises = _exerciseRepository.GetAll().ToList();
-			int id = ExerciseController.GetIdOption(exercises);
-			var exercise = _exerciseRepository.GetExerciseById(id);
-			ExerciseController.PrintExercise(exercise);
+			if (exercises.Count == 0)
+			{
+				Console.WriteLine("The database is empty");
+				Console.ReadKey();
+			}
+			else
+			{
+				int id = ExerciseController.GetIdOption(exercises);
+				var exercise = _exerciseRepository.GetExerciseById(id);
+				ExerciseController.PrintExercise(exercise);
+			}
 		}
 
 		internal void UpdateExercise()
 		{
 			var exercises = _exerciseRepository.GetAll().ToList();
-			int id = ExerciseController.GetIdOption(exercises);
-			var exercise = _exerciseRepository.GetExerciseById(id);
-			var updatedExercise = ExerciseController.GetUpdateInfo(exercise);
-			_exerciseRepository.Update(updatedExercise);
+			if (exercises.Count == 0)
+			{
+				Console.WriteLine("The database is empty");
+				Console.ReadKey();
+			}
+			else
+			{
+				int id = ExerciseController.GetIdOption(exercises);
+				var exercise = _exerciseRepository.GetExerciseById(id);
+				var updatedExercise = ExerciseController.GetUpdateInfo(exercise);
+				_exerciseRepository.Update(updatedExercise);
+			}
 		}
 	}
 }
